requirements.txt
adal==1.2.7
appnope==0.1.4
argcomplete==3.3.0
asttokens==2.4.1
azure-common==1.1.28
azure-core==1.30.1
azure-graphrbac==0.61.1
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-network==25.3.0
azure-mgmt-resource==23.1.1
azure-mgmt-storage==21.1.0
azureml==0.2.7
azureml-core==1.56.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.1.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
contextlib2==21.6.0
cryptography==42.0.7
debugpy==1.8.1
decorator==5.1.1
docker==7.0.0
exceptiongroup==1.2.1
executing==2.0.1
humanfriendly==10.0
idna==3.7
importlib-metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
isodate==0.6.1
jedi==0.19.1
jeepney==0.8.0
jmespath==1.0.1
joblib==1.4.2
jsonpickle==3.0.4
jupyter-client==8.6.1
jupyter-core==5.7.2
knack==0.11.0
matplotlib-inline==0.1.7
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
msrestazure==0.6.4
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
paramiko==3.4.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.2
portalocker==2.8.2
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.22
pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.31.0
requests-oauthlib==2.0.0
scipy==1.13.0
SecretStorage==3.3.3
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
tornado==6.4
traitlets==5.14.3
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.18.2
